Don't sweep the GitHub integration
This is a static IdP which has been manually setup to support testing
the Access Group integrations. This commit ensures we don't
automatically remove it.

+++ cloudflare/resource_cloudflare_access_identity_provider_test.go
@@ -34,6 +34,12 @@ func testSweepCloudflareAccessIdentityProviders(r string) error {
 	}
 
 	for _, idp := range accessIDPs {
+		// e75d3adf-964b-4b15-acff-e7e20838c33c is a static GitHub integration used
+		// in acceptance tests so we don't want to remove it.
+		if idp.ID == "e75d3adf-964b-4b15-acff-e7e20838c33c" {
+			continue
+		}
+
 		log.Printf("[INFO] Deleting Access Identity Provider ID: %s", idp.ID)
 		_, err := client.DeleteAccessIdentityProvider(accountID, idp.ID)
